Tests/LibWeb/Text/expected/css/abspos-inline-static-position-after-wrapped-inline.txt

target x: 20
target y: 20
